// src/graphql/cloudshelf/generated/cloudshelf_types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  GlobalId: { input: any; output: any; }
  Latitude: { input: any; output: any; }
  Longitude: { input: any; output: any; }
  UTCDateTime: { input: any; output: any; }
};
// ...
export type CloudshelfEnginePayload = {
  __typename?: 'CloudshelfEnginePayload';
  cloudshelf?: Maybe<Cloudshelf>;
  device?: Maybe<Device>;
  engineType: EngineType;
  errorMessage?: Maybe<Scalars['String']['output']>;
  handoffPayload?: Maybe<HandoffPayload>;
  inMaintenanceMode: Scalars['Boolean']['output'];
  owningOrganisation?: Maybe<Organisation>;
  plugins: Array<Plugin>;
  salesAssistants: Array<SalesAssistant>;
  status: CloudshelfPayloadStatus;
};
// ...
export type CloudshelfPaginatedPayload = {
  __typename?: 'CloudshelfPaginatedPayload';
  edges?: Maybe<Array<CloudshelfEdge>>;
  pageInfo?: Maybe<CloudshelfPageInfo>;
  /** The total number of items available */
  totalCount: Scalars['Int']['output'];
};
// ...
export enum CurrencyCode {
  /** United Arab Emirates Dirham */
  Aed = 'AED',
  /** Australian Dollar */
  Aud = 'AUD',
  /** Canadian Dollar */
  Cad = 'CAD',
  /** Swiss Franc */
  Chf = 'CHF',
  /** Chilean Peso */
  Clp = 'CLP',
  /** Euro */
  Eur = 'EUR',
  /** Great British Pound */
  Gbp = 'GBP',
  /** Hungarian Forint */
  Huf = 'HUF',
  /** Israeli New Shekel */
  Ils = 'ILS',
  /** Peruvian Sol */
  Pen = 'PEN',
  /** Saudi Arabian Riyal */
  Sar = 'SAR',
  Unknown = 'UNKNOWN',
  /** United States Dollar */
  Usd = 'USD'
}
// ...
export type IncludablePdpBlock = {
  __typename?: 'IncludablePDPBlock';
  metadataKey?: Maybe<Scalars['String']['output']>;
  productDataType?: Maybe<PdpProductDataType>;
  type: PdpBlockType;
};

export type IngestionStatsPayload = {
  __typename?: 'IngestionStatsPayload';
  imageCount: Scalars['Int']['output'];
  productCount: Scalars['Int']['output'];
  productGroupCount: Scalars['Int']['output'];
  productGroupWithValidProductCount: Scalars['Int']['output'];
  validProductCount: Scalars['Int']['output'];
  variantCount: Scalars['Int']['output'];
};
// ...
export type Order = {
  __typename?: 'Order';
  /** The date and time this entity was created. */
  createdAt: Scalars['UTCDateTime']['output'];
  discountCode?: Maybe<Scalars['String']['output']>;
  /** A unique internal GlobalId for this entity. */
  id: Scalars['GlobalId']['output'];
  lines: Array<OrderLine>;
  owningOrganisation: Organisation;
  session?: Maybe<Session>;
  status: OrderStatus;
  thirdPartyId?: Maybe<Scalars['String']['output']>;
  /** The date and time this entity was last updated. */
  updatedAt: Scalars['UTCDateTime']['output'];
};

export type OrderEdge = {
  __typename?: 'OrderEdge';
  /** The cursor for provided node to be used in pagination */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The Order entity */
  node?: Maybe<Order>;
};

export type OrderInput = {
  /** The discount code that was used on this order */
  discountCode?: InputMaybe<Scalars['String']['input']>;
  /** Use this field to provide either a Cloudshelf gid, or your own external gid. If the external gid already exists, the existing record will be updated. If the external gid does not exist, a new record will be created. */
  id?: InputMaybe<Scalars['GlobalId']['input']>;
  /** An array of lines that make up this order */
  lines?: InputMaybe<Array<OrderLineInput>>;
  /** Use this field to provide a third party id to this order */
  newThirdPartyId?: InputMaybe<Scalars['String']['input']>;
  /** The GlobalID of the session this order was created in */
  sessionId?: InputMaybe<Scalars['GlobalId']['input']>;
  status?: InputMaybe<OrderStatus>;
  /** Use this field to provide a third party id to this order */
  thirdPartyId?: InputMaybe<Scalars['String']['input']>;
};

export type OrderLine = {
  __typename?: 'OrderLine';
  /** The date and time this entity was created. */
  createdAt: Scalars['UTCDateTime']['output'];
  currencyCode: CurrencyCode;
  /** A unique internal GlobalId for this entity. */
  id: Scalars['GlobalId']['output'];
  order: Order;
  platformProvidedId?: Maybe<Scalars['GlobalId']['output']>;
  platformProvidedProductId?: Maybe<Scalars['GlobalId']['output']>;
  platformProvidedProductVariantId?: Maybe<Scalars['GlobalId']['output']>;
  price: Scalars['Float']['output'];
  productId: Scalars['String']['output'];
  productName: Scalars['String']['output'];
  productVariantId: Scalars['String']['output'];
  productVariantName: Scalars['String']['output'];
  quantity: Scalars['Int']['output'];
  /** The date and time this entity was last updated. */
  updatedAt: Scalars['UTCDateTime']['output'];
};
// ...
export enum OrderStatus {
  /** The order is a draft basket */
  DraftBasket = 'DRAFT_BASKET',
  Paid = 'PAID',
  /** The order has been partially refunded */
  PartiallyRefunded = 'PARTIALLY_REFUNDED',
  /** The order has been placed & paid for */
  Placed = 'PLACED',
  /** The order has been fully refunded */
  Refunded = 'REFUNDED'
}

export type OrderUpsertPayload = {
  __typename?: 'OrderUpsertPayload';
  /** An array of Orders that were created or updated */
  orders: Array<Order>;
  /** An array of errors that occurred during the upsert operation */
  userErrors: Array<UserError>;
};
// ...
export type Query = {
  __typename?: 'Query';
  allIngestionStats: Array<IngestionStatsPayload>;
  /** Returns an array of PDP Blocks that are available to add to the PDP */
  availablePDPBlocks: Array<IncludablePdpBlock>;
  /** Returns a boolean value which indicates if the organisation can register a new device */
  canRegisterDevice: Scalars['Boolean']['output'];
  /** Returns a Cloudshelf entity */
  cloudshelf?: Maybe<Cloudshelf>;
  cloudshelfEnginePayload: CloudshelfEnginePayload;
  /** Returns a paginated array of Cloudshelves */
  cloudshelves: CloudshelfPaginatedPayload;
  /** An internal function for getting a Cloudshelf Authentication Token. */
  customToken?: Maybe<Scalars['String']['output']>;
  /** Returns a Device entity */
  device?: Maybe<Device>;
  /** Returns a paginated array of Devices. */
  devices: DevicePaginatedPayload;
  engineProducts: EngineProductWithAdditionalInfoPayload;
  /** Returns a Filter entity */
  filter?: Maybe<Filter>;
  getMobileHandoff?: Maybe<MobileHandoff>;
  getPresignedCustomisationUrl: Scalars['String']['output'];
  getVersionByType: KnownVersion;
  /** Returns all filters that can possibly be included in the Cloudshelf */
  includeableFilters: Array<CloudshelfIncludableFilter>;
  ingestionStats: IngestionStatsPayload;
  isInMaintenanceMode: Scalars['Boolean']['output'];
  isNobleEnabled: Scalars['Boolean']['output'];
  isTrackedURLScanned: Scalars['Boolean']['output'];
  /** Returns a location entity */
  location?: Maybe<Location>;
  /** Returns a paginated array of locations */
  locations: LocationPaginatedPayload;
  /** Returns the currently authenticated user. */
  me: User;
  nobleQueueByTaskType: TaskQueue;
  nobleQueues: Array<TaskQueue>;
  nobleTask?: Maybe<NobleTask>;
  nobleTasksByTypes: NobleTaskResponse;
  /** Returns an Order entity. */
  order?: Maybe<Order>;
  /** Returns a paginated array of Order entities */
  orders: OrderPaginatedPayload;
  /** Returns an Organisation entity */
  organisation?: Maybe<Organisation>;
  /** Returns if the install has been completed for an organisation */
  organisationInstallComplete?: Maybe<Scalars['Boolean']['output']>;
  /** Returns a paginated array of user access rights for the current organisation */
  organisationUsers: UserForOrganisationPaginatedPayload;
  /** Returns a paginated array of organisations */
  organisations: OrganisationPaginatedPayload;
  pausedNobleQueues: Array<NobleTaskType>;
  /** Returns a paginated array of plugins */
  plugins: PluginPaginatedPayload;
  /** Returns a Product entity. */
  product?: Maybe<Product>;
  /** Returns a Product Group */
  productGroup?: Maybe<ProductGroup>;
  /** Returns a paginated array of product groups */
  productGroups: ProductGroupPaginatedPayload;
  /** Returns a paginated array of Product entities. */
  products: ProductPaginatedPayload;
  /** Returns public information about a device */
  publicDevice?: Maybe<PublicDevicePayload>;
  /** Returns a sales assistant entity */
  salesAssistant?: Maybe<SalesAssistant>;
  /** Returns a paginated array of Sales Assistants */
  salesAssistants: SalesAssistantPaginatedPayload;
  sessions: Array<Session>;
  /** Returns a subscription entity */
  subscription?: Maybe<SubscriptionRecord>;
  /** This is an internal function. This function is not intended to be included in the final release. Only exists due to CS-1273 */
  subscriptionCurrentOrg?: Maybe<SubscriptionRecord>;
  /** Returns a list of currently available subscription plans */
  subscriptionPlans?: Maybe<Array<SubscriptionPlan>>;
  syncStats: SyncStatsPayload;
  testFilters: Scalars['Boolean']['output'];
  /** Returns a theme entity */
  theme?: Maybe<Theme>;
  /** Returns a paginated array of Themes */
  themes: ThemePaginatedPayload;
  usageStatsSessions: UsageSessionsResponse;
  /** Returns a User entity */
  user: User;
  /** Query a webhook by ID */
  webhook?: Maybe<Webhook>;
  /** Query the list of registered webhooks */
  webhooks: WebhookPaginatedPayload;
};
// ...
export type UserError = {
  __typename?: 'UserError';
  /** An error code that can be used to programmatically identify the error */
  code: UserErrorCode;
  /** This field provides more explicit information about the error */
  message: Scalars['String']['output'];
};

/** A high level error code which is used to indicate the type of error that has occurred. For more information see the message field. */
export enum UserErrorCode {
  ActionDisallowed = 'ACTION_DISALLOWED',
  /** The data provided for the given upsert function was missing and is required for entity creation */
  EntityCreationMissingField = 'ENTITY_CREATION_MISSING_FIELD',
  /** The data provided for the given upsert function was invalid for entity creation or updating */
  EntityInvalidField = 'ENTITY_INVALID_FIELD',
  EntityInUse = 'ENTITY_IN_USE',
  EntityNotFound = 'ENTITY_NOT_FOUND',
  /** An error occurred while attempting to upload an image */
  ImageUploadError = 'IMAGE_UPLOAD_ERROR',
  InvalidArgument = 'INVALID_ARGUMENT',
  /** The file provided was invalid, or could not be accessed. */
  InvalidFile = 'INVALID_FILE',
  InvalidHmac = 'INVALID_HMAC',
  TemporaryRestriction = 'TEMPORARY_RESTRICTION',
  UnknownError = 'UNKNOWN_ERROR'
}
// ...
export const GetOrderDocument = gql`
    query GetOrder($orderId: GlobalId!) {
  order(id: $orderId) {
    id
    thirdPartyId
    createdAt
    updatedAt
    status
    lines {
      id
      platformProvidedId
      platformProvidedProductId
      platformProvidedProductVariantId
      productId
      productName
      productVariantId
      productVariantName
      quantity
      price
      currencyCode
    }
  }
}
    `;
export const UpsertOrdersDocument = gql`
    mutation UpsertOrders($input: [OrderInput!]!) {
  upsertOrders(input: $input) {
    orders {
      id
      status
    }
    userErrors {
      code
      message
    }
  }
}
    `;
export type GetOrderQueryVariables = Exact<{
  orderId: Scalars['GlobalId']['input'];
}>;


export type GetOrderQuery = { __typename?: 'Query', order?: { __typename?: 'Order', id: any, thirdPartyId?: string | null, createdAt: any, updatedAt: any, status: OrderStatus, lines: Array<{ __typename?: 'OrderLine', id: any, platformProvidedId?: any | null, platformProvidedProductId?: any | null, platformProvidedProductVariantId?: any | null, productId: string, productName: string, productVariantId: string, productVariantName: string, quantity: number, price: number, currencyCode: CurrencyCode }> } | null };

export type UpsertOrdersMutationVariables = Exact<{
  input: Array<OrderInput> | OrderInput;
}>;


export type UpsertOrdersMutation = { __typename?: 'Mutation', upsertOrders: { __typename?: 'OrderUpsertPayload', orders: Array<{ __typename?: 'Order', id: any, status: OrderStatus }>, userErrors: Array<{ __typename?: 'UserError', code: UserErrorCode, message: string }> } };